requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
openai>=1.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
spacy>=3.6.0
python-docx>=0.8.11
PyPDF2>=3.0.1
plotly>=5.15.0
jupyter>=1.0.0
pdfplumber>=0.11.0
python-dotenv>=1.0.1